pyproject.toml
[project]
name = "vimlayer"
version = "0.0.0"
description = "Vim-style UI element navigation for macOS"
requires-python = ">=3.13"
dependencies = [
    "pyobjc-framework-Cocoa; sys_platform == 'darwin'",
    "pyobjc-framework-Quartz; sys_platform == 'darwin'",
    "pyobjc-framework-ApplicationServices; sys_platform == 'darwin'",
    "pyobjc>=12.1; sys_platform == 'darwin'",
    "PyQt6; sys_platform == 'linux'",
    "python-xlib; sys_platform == 'linux'",
    "ewmh; sys_platform == 'linux'",
    "pynput; sys_platform == 'linux'",
    "PyGObject; sys_platform == 'linux'",
]

[dependency-groups]
dev = [
    "py2app",
    "setuptools",
    "ruff",
    "pytest",
    "pytest-mock",
    "mypy",
    "pre-commit",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint.per-file-ignores]
"vimlayer/main.py" = ["E402"]

[tool.mypy]
python_version = "3.13"
check_untyped_defs = true
ignore_missing_imports = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_untyped_calls = false
disallow_untyped_defs = false
disable_error_code = [
    "union-attr",
    "arg-type",
    "dict-item",
    "name-defined",
    "var-annotated",
    "misc",
    "call-overload",
    "attr-defined",
    "no-any-return",
]

[project.scripts]
vimlayer = "vimlayer.main:main"

[tool.py2app]
app = ["vimlayer/main.py"]
argv_emulation = false
packages = [
    "objc",
    "AppKit",
    "Foundation",
    "Quartz",
    "CoreFoundation",
    "CoreText",
    "ApplicationServices",
    "HIServices",
    "PyObjCTools",
]

[tool.py2app.plist]
CFBundleName = "VimLayer"
CFBundleIdentifier = "com.vimlayer.app"
CFBundleVersion = "0.0.0"
CFBundleShortVersionString = "0.0.0"
LSUIElement = true
NSAccessibilityUsageDescription = "VimLayer needs Accessibility access to detect UI elements and simulate clicks."
